+++ app/src/main/java/org/simple/clinic/home/overdue/search/OverdueSearchUpdate.kt
@@ -3,6 +3,10 @@ package org.simple.clinic.home.overdue.search
 import com.spotify.mobius.Next
 import com.spotify.mobius.Next.noChange
 import com.spotify.mobius.Update
+import org.simple.clinic.home.overdue.search.OverdueSearchQueryValidator.Result.Empty
+import org.simple.clinic.home.overdue.search.OverdueSearchQueryValidator.Result.LengthTooShort
+import org.simple.clinic.home.overdue.search.OverdueSearchQueryValidator.Result.Valid
+import org.simple.clinic.mobius.dispatch
 import org.simple.clinic.mobius.next
 
 class OverdueSearchUpdate : Update<OverdueSearchModel, OverdueSearchEvent, OverdueSearchEffect> {
@@ -14,7 +18,16 @@ class OverdueSearchUpdate : Update<OverdueSearchModel, OverdueSearchEvent, Overd
           model.overdueSearchQueryChanged(event.searchQuery),
           ValidateOverdueSearchQuery(event.searchQuery)
       )
-      is OverdueSearchQueryValidated -> noChange()
+      is OverdueSearchQueryValidated -> searchQueryValidated(event)
+    }
+  }
+
+  private fun searchQueryValidated(event: OverdueSearchQueryValidated): Next<OverdueSearchModel, OverdueSearchEffect> {
+    return when (val result = event.result) {
+      // TODO: Start overdue search
+      is Valid -> dispatch(AddQueryToOverdueSearchHistory(result.searchQuery))
+      Empty,
+      LengthTooShort -> noChange()
     }
   }
 }