+++ app/src/test/java/org/simple/clinic/home/overdue/search/OverdueSearchUpdateTest.kt
@@ -3,9 +3,11 @@ package org.simple.clinic.home.overdue.search
 import com.spotify.mobius.test.NextMatchers.hasEffects
 import com.spotify.mobius.test.NextMatchers.hasModel
 import com.spotify.mobius.test.NextMatchers.hasNoEffects
+import com.spotify.mobius.test.NextMatchers.hasNoModel
 import com.spotify.mobius.test.UpdateSpec
 import com.spotify.mobius.test.UpdateSpec.assertThatNext
 import org.junit.Test
+import org.simple.clinic.home.overdue.search.OverdueSearchQueryValidator.Result.Valid
 
 class OverdueSearchUpdateTest {
 
@@ -37,6 +39,18 @@ class OverdueSearchUpdateTest {
         .then(assertThatNext(
             hasModel(defaultModel.overdueSearchQueryChanged(searchQuery)),
             hasEffects(ValidateOverdueSearchQuery(searchQuery))
+        ))
+  }
+
+  @Test
+  fun `when search query is validated and is valid, then update model and add query to search history`() {
+    val searchQuery = "Babri"
+    updateSpec
+        .given(defaultModel)
+        .whenEvent(OverdueSearchQueryValidated(Valid(searchQuery)))
+        .then(assertThatNext(
+            hasNoModel(),
+            hasEffects(AddQueryToOverdueSearchHistory(searchQuery))
         ))
   }
 }